+++ home/guide/views/examples/tab/tab-default.vue
@@ -3,95 +3,93 @@
     <div class="button-outer">
       <ev-button
         :type="'primary'"
-        :text="'Tab Add'"
         @click="addTab"
-      />
+      >
+        Tab Add
+      </ev-button>
     </div>
     <div class="tab-default-outer">
       <ev-tabs
-        :tab-data="list"
+        v-model="activeTabName"
+        :use-tab-moving="true"
+        @remove-tab="removeTab"
+        @change-tab="changeTab"
       >
-        <ev-tab
-          :tab-prop="tabItem"
-        />
-        <ev-tab
+        <ev-tab-panel
           v-for="item in tabItems"
-          :key="item.id"
-          :tab-prop="item"
-        />
+          :key="item.value"
+          :title="item.title"
+          :value="item.value"
+        >
+          <component
+            :is="item.content"
+          />
+        </ev-tab-panel>
       </ev-tabs>
     </div>
   </div>
 </template>
+
 <script>
   import targetComponent1 from '../checkbox/checkbox-group-list-handling';
   import targetComponent2 from '../table/table-buffer';
   import targetComponent3 from '../chart/chart.bar.stack';
   import targetComponent4 from '../table/table-page';
 
-  const componentList = [targetComponent1, targetComponent2, targetComponent3];
-
   export default {
+    components: {
+      targetComponent1,
+      targetComponent2,
+      targetComponent3,
+      targetComponent4,
+    },
     data() {
       return {
-        scroll: false,
-        tabItem: {
-          title: 'appended tab',
-          isActive: false,
-          targetComponent: {
-            keyName: 'targetComponent1',
-            component: targetComponent1,
-          },
-        },
-        tabItems: [{
-            title: 'appended tab',
-            isActive: false,
-            targetComponent: {
-              keyName: 'targetComponent3',
-              component: targetComponent3,
-            },
-          },
-          {
-            title: 'appended tab',
-            isActive: false,
-            targetComponent: {
-              keyName: 'targetComponent4',
-              component: targetComponent4,
-            },
-          },
-        ],
-        list: this.createList(),
+        seq: 1,
+        tabItems: [],
+        activeTabName: '1',
       };
     },
+    created() {
+      this.createTabs();
+    },
     methods: {
-      addTab() {
-        this.list.push({
-          title: 'appended tab',
-          isActive: false,
-          targetComponent: {
-            keyName: 'target-component1',
-            component: componentList[0],
-          },
-        });
-      },
-      createList() {
-        const result = [];
-        for (let i = 0, length = 3; i < length; i++) {
-          result.push({
-            title: 'appended tab',
-            isActive: false,
-            width: 100,
-            targetComponent: {
-              keyName: `target-component${i + 1}`,
-              component: componentList[i],
-            },
+      createTabs() {
+        for (let ix = 0; ix < 10; ix++) {
+          this.tabItems.push({
+            title: `appended tab${this.seq}`,
+            value: `${this.seq}`,
+            content: `targetComponent${(this.seq % 4) + 1}`,
           });
+
+          this.seq++;
         }
-        return result;
+      },
+      addTab() {
+        this.tabItems.push({
+          title: `appended tab${this.seq}`,
+          value: `${this.seq}`,
+          content: `targetComponent${(this.seq % 4) + 1}`,
+        });
+        this.activeTabName = `${this.seq}`;
+        this.seq++;
+      },
+      removeTab(value) {
+        for (let ix = 0; ix < this.tabItems.length; ix++) {
+          if (this.tabItems[ix].value === value) {
+            this.tabItems.splice(ix, 1);
+            break;
+          }
+        }
+      },
+      changeTab(oldValue, newValue) {
+        window.console.log(`oldValue: ${oldValue}`);
+        window.console.log(`newValue: ${newValue}`);
       },
     },
   };
 </script>
+
 <style scoped>
   p{
     margin-left: 30px;
@@ -103,7 +101,7 @@
     width: 100%;
   }
   .tab-default-outer {
-    width: 1500px;
-    height: 1000px;
+    width: 100%;
+    height: 100%;
   }
 </style>
